[lgdo] rename source file
