+++ src/pygama/lgdo/__init__.py
@@ -37,6 +37,7 @@ browsed easily in python like any `HDF5 <https://www.hdfgroup.org>`_ file using
 
 from pygama.lgdo.array import Array
 from pygama.lgdo.arrayofequalsizedarrays import ArrayOfEqualSizedArrays
+from pygama.lgdo.encoded import VectorOfEncodedVectors
 from pygama.lgdo.fixedsizearray import FixedSizeArray
 from pygama.lgdo.lgdo import LGDO
 from pygama.lgdo.lgdo_utils import copy
@@ -44,7 +45,6 @@ from pygama.lgdo.lh5_store import LH5Iterator, LH5Store, load_dfs, load_nda, ls,
 from pygama.lgdo.scalar import Scalar
 from pygama.lgdo.struct import Struct
 from pygama.lgdo.table import Table
-from pygama.lgdo.vectorofencodedvectors import VectorOfEncodedVectors
 from pygama.lgdo.vectorofvectors import (
     VectorOfVectors,
     build_cl,
