+++ src/pygama/lgdo/encoded.py
