+++ src/pygama/lgdo/waveform_table.py
@@ -12,8 +12,8 @@ import numpy as np
 
 from pygama.lgdo.array import Array
 from pygama.lgdo.arrayofequalsizedarrays import ArrayOfEqualSizedArrays
+from pygama.lgdo.encoded import VectorOfEncodedVectors
 from pygama.lgdo.table import Table
-from pygama.lgdo.vectorofencodedvectors import VectorOfEncodedVectors
 from pygama.lgdo.vectorofvectors import VectorOfVectors
 
 log = logging.getLogger(__name__)
